+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed 
+- Update dependency io.cucumber:messages up to v30
+- Write test cases in canonical order ([#98](https://github.com/cucumber/testng-xml-formatter/pull/98))
 
 ## [0.11.0] - 2025-11-14
 ### Added